moved lifeshard stuff into a separate directory
