+++ src/main/java/teamport/aether/mixin/item/lifeshard/PlayerListMixinKeepEHPonDeath.java
@@ -1,4 +1,4 @@
-package teamport.aether.mixin.item;
+package teamport.aether.mixin.item.lifeshard;
 
 import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.server.entity.player.PlayerServer;
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import teamport.aether.api.HealthHelper;
 
 @Mixin(value = PlayerList.class)
-public class PlayerListMixin {
+public class PlayerListMixinKeepEHPonDeath {
 
     @Inject(method = "recreatePlayerEntity", at = @At(value = "INVOKE", target = "Lnet/minecraft/core/player/inventory/container/ContainerInventory;transferAllContents(Lnet/minecraft/core/player/inventory/container/ContainerInventory;)V"), remap = false)
     public void keepExtraHealthServer(final PlayerServer previousPlayer, final int i, final CallbackInfoReturnable<PlayerServer> cir, @Local(name = "newPlayer") final PlayerServer newPlayer) {
