+++ src/main/java/teamport/aether/mixin/item/lifeshard/PlayerMixinExtraHealth.java
@@ -1,4 +1,4 @@
-package teamport.aether.mixin.item;
+package teamport.aether.mixin.item.lifeshard;
 
 import com.mojang.nbt.tags.CompoundTag;
 import net.minecraft.core.entity.player.Player;
@@ -11,7 +11,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import teamport.aether.api.IVariableHealthPlayer;
 
 @Mixin(value = Player.class, remap = false)
-public class LifeShardExtraHealth
+public class PlayerMixinExtraHealth
         implements IVariableHealthPlayer {
     @Unique
     public int extraHealth = 0;
